Underwater-Visual-Perception-System/requirements.txt
opencv-python-headless
ultralytics
numpy==1.26.4
matplotlib
torch
torchvision
streamlit
pandas
Pillow
